+++ frontend/src/main.jsx
@@ -6,48 +6,8 @@ import ReactDOM from "react-dom/client";
 import "@/style/index.css";
 
 // router
-import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import App from "@/App.jsx";
-import { Home } from "@/pages/home/Home.jsx";
-import { Horoscope } from "@/pages/horoscope/Horoscope.jsx";
-import { ContactUs } from "@/pages/contact-us/ContactUs.jsx";
-import { ProductsList } from "@/pages/products-list/ProductsList.jsx";
-import { ProductDetail } from "@/pages/product-detail/ProductDetail.jsx";
-import DevDocument from "./pages/DevDocument";
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <App />,
-    children: [
-      {
-        path: "/",
-        element: <Home />,
-      },
-      {
-        path: "/horoscope",
-        element: <Horoscope />,
-      },
-      {
-        path: "/contact-us",
-        element: <ContactUs />,
-      },
-      {
-        path: "/products-list",
-        element: <ProductsList />,
-      },
-      {
-        path: "/product-detail",
-        element: <ProductDetail />,
-      },
-    ],
-  },
-  // Dev Document
-  {
-    path: "/document",
-    element: <DevDocument />,
-  }
-]);
+import { RouterProvider } from "react-router-dom";
+import router from "@/router/rounter";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
